requirements.txt
pydub
argparse
wave
hashlib
numpy
pycryptodome